QQ一键加群、去除新浪微博分享，增加QQ空间分享

+++ MCHSecretary/AppInfoTableView.m
@@ -246,7 +246,7 @@
 -(void) clickItem:(NSInteger)appid{
     NSLog(@"clickItem : %ld", (long)appid);
     if(_delegate){
-        [_delegate showAppInfo: [NSString stringWithFormat:@"%ld", appid]];
+        [_delegate showAppInfo: [NSString stringWithFormat:@"%ld", (long)appid]];
     }
 }
